Fixed some stuff, added gitignore to reduce file bloat.
