+++ .gitignore
@@ -1,0 +1,3 @@
+# Class files #
+###############
+.class
